modules/zulip-crawler/requirements.txt
httpx>=0.27.0
anthropic>=0.25.0
python-dotenv>=1.0.0
